requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
appdirs==1.4.4
apturl==0.5.2
attrs==26.1.0
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
build==1.4.2
buildozer==1.5.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
Cython==3.0.8
dbus-python==1.3.2
defer==1.0.6
discord.py==2.7.1
distlib==0.4.0
distro==1.9.0
distro-info==1.7+build1
duplicity==2.1.4
fasteners==0.18
filelock==3.25.2
frozenlist==1.8.0
gbinder-python==1.3.1
html5lib==1.1
httplib2==0.20.4
icoextract==0.1.6
idna==3.6
inquirerpy==0.3.4
Jinja2==3.1.6
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
multidict==6.7.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==26.0
paramiko==2.12.0
pefile==2023.2.7
pexpect==4.9.0
pfzy==0.3.4
pillow==10.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyproject_hooks==1.2.0
pystyle==2.9
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-discovery==1.2.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
setproctitle==1.3.3
setuptools==68.1.2
sh==1.14.3
six==1.16.0
soupsieve==2.5
tldextract==3.1.2
toml==0.10.2
tqdm==4.67.3
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
virtualenv==21.2.0
wadllib==1.3.6
wcwidth==0.6.0
webencodings==0.5.1
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.23.0
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5
